+++ app/support/ExtAuth/OAuth2Adapter.ts
@@ -218,7 +218,7 @@ type GeneralErrorResponse = {
   error: string | { message: string };
 };
 
-async function fetchJSON<R extends object>(url: string, params?: NodeJS.fetch.RequestInit) {
+async function fetchJSON<R extends object>(url: string, params?: RequestInit) {
   const response = await fetch(url, params);
   let respBody: R | GeneralErrorResponse = {
     error: response.ok ? 'Unknown error' : `HTTP error ${response.status}`,