+++ migrations/20160405133637_initial.js
@@ -134,7 +134,6 @@ exports.up = function (knex) {
       table.uuid('uid').defaultTo(knex.raw('gen_random_uuid()')).notNullable().unique();
       table.text('body');
       table.uuid('post_id').notNullable();
-      // TODO: foreign key post_id (fix tests)
       table
         .uuid('user_id')
         .notNullable()
@@ -156,7 +155,6 @@ exports.up = function (knex) {
       table.uuid('uid').defaultTo(knex.raw('gen_random_uuid()')).notNullable().unique();
       table.text('name').notNullable();
       table.uuid('user_id').notNullable();
-      // TODO: foreign key user_id (fix tests)
       table.timestamp('created_at').defaultTo(knex.fn.now()).notNullable();
       table.timestamp('updated_at').defaultTo(knex.fn.now()).notNullable();
 
@@ -175,14 +173,14 @@ exports.up = function (knex) {
         .references('uid')
         .inTable('users')
         .onUpdate('cascade')
-        .onDelete('cascade'); // Todo: add constraint(type=='group')
+        .onDelete('cascade');
       table
         .uuid('user_id')
         .notNullable()
         .references('uid')
         .inTable('users')
         .onUpdate('cascade')
-        .onDelete('cascade'); // Todo: add constraint(type=='user')
+        .onDelete('cascade');
       table.timestamp('created_at').defaultTo(knex.fn.now()).notNullable();
 
       table.index('created_at', 'group_admins_created_at_idx', 'btree');
@@ -192,7 +190,6 @@ exports.up = function (knex) {
     knex.schema.createTable('likes', function (table) {
       table.increments().notNullable().primary();
       table.uuid('post_id').notNullable();
-      // TODO: foreign key post_id (fix tests)
       table
         .uuid('user_id')
         .notNullable()